[Refactor] 기존 코드 별도의 폴더로 이동
