+++ hunmin/src/main/java/com/nhnacademy/legacy/Breakout.java
@@ -1,14 +1,10 @@
-package com.nhnacademy;
+package com.nhnacademy.legacy;
 
 import javafx.animation.AnimationTimer;
 import javafx.application.Application;
-import javafx.application.Platform;
 import javafx.scene.Scene;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.ButtonType;
 import javafx.scene.input.KeyCode;
 import javafx.scene.layout.StackPane;
 import javafx.scene.paint.Color;
@@ -17,6 +13,9 @@ import javafx.stage.Stage;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * 메인 BreakOut
+ */
 public class Breakout extends Application {
     private boolean moveLeft = false;
     private boolean moveRight = false;
